Move constants and static classes out of HivePageSource
Make BucketAdapter and BucketValidator top level classes
