+++ plugin/trino-hive/src/main/java/io/trino/plugin/hive/MergeFileWriter.java
@@ -42,9 +42,9 @@ import static com.google.common.base.Preconditions.checkArgument;
 import static com.google.common.collect.ImmutableList.toImmutableList;
 import static io.trino.orc.OrcWriter.OrcOperation.DELETE;
 import static io.trino.orc.OrcWriter.OrcOperation.INSERT;
-import static io.trino.plugin.hive.HivePageSource.BUCKET_CHANNEL;
-import static io.trino.plugin.hive.HivePageSource.ORIGINAL_TRANSACTION_CHANNEL;
-import static io.trino.plugin.hive.HivePageSource.ROW_ID_CHANNEL;
+import static io.trino.plugin.hive.HivePageSourceProvider.BUCKET_CHANNEL;
+import static io.trino.plugin.hive.HivePageSourceProvider.ORIGINAL_TRANSACTION_CHANNEL;
+import static io.trino.plugin.hive.HivePageSourceProvider.ROW_ID_CHANNEL;
 import static io.trino.plugin.hive.HiveStorageFormat.ORC;
 import static io.trino.plugin.hive.acid.AcidSchema.ACID_COLUMN_NAMES;
 import static io.trino.plugin.hive.acid.AcidSchema.createAcidSchema;
